✅🤖 give the mdim e2e tests their own worker and wrangler config
Restores grapher-config.e2e.worker.ts to its original scope (R2 config
endpoints) and moves the meta-tag-rewriting harness — the seeded
in-memory ASSETS stub and the /__test__/rewrite-meta-tags, seed-asset
and clear-assets routes — into mdim-meta-tags.e2e.worker.ts, mirroring
how the search e2e tests are set up.
